Collect Firewall resource from other modules
When all Firewall resources in other Puppet-Finland modules are made virtual
resources containing the "default" tag, packet filtering can be turned on and
off in a declaration order independent way by including, or not including the
packetfilter module. Also, when $manage_packetfilter parameters are present in
the other modules, packet filtering can be selectively turned off on a
per-service basis on each node.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>